pyproject.toml
[project]
name = "polarbadge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.4"
dependencies = [
    "click>=8.3.0",
    "jinja2>=3.1.6",
    "opencv-python>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "plutoprint>=0.11.0",
    "pydantic>=2.11.9",
    "requests>=2.32.5",
    "treepoem>=3.28.0",
    "weasyprint>=66.0",
]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
]
